Fixup invalid configuration error message
* Fix invalid parameter used in error message
* Re-organize test code
